feat: add rist tabs in pool modal
