+++ src/components/PoolModal/BondMode/index.ts
@@ -1,0 +1,1 @@
+export { default } from './BondMode';
