+++ src/components/PoolModal/DepositTab/DepositTab.module.scss
@@ -1,0 +1,42 @@
+.btn {
+  width: 100%;
+  padding: 0;
+  font: var(--font-body-small);
+  justify-content: center;
+  height: 48px;
+}
+
+.slider {
+  padding: 15px 5px 64px;
+}
+
+.info {
+  margin-bottom: 8px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  text-transform: uppercase;
+}
+
+.infoTitle {
+  color: var(--secondary-text);
+  font: var(--font-text-6);
+}
+
+.infoValue {
+  color: var(--primary-text);
+  font: var(--font-text-8);
+}
+
+.input {
+  margin-bottom: 10px;
+}
+
+.errors {
+  height: 16px;
+  font: var(--font-text-6);
+  color: var(--red-error-color);
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-end;
+}
