+++ src/components/PoolModal/WithdrawTab/index.ts
@@ -1,0 +1,1 @@
+export { default } from './WithdrawTab';
