+++ src/componentsNew/SlippageModal/SlippageModal.module.scss
