+++ src/componentsNew/SwapModal/index.ts
@@ -1,0 +1,2 @@
+export { default } from './SwapForm';
+export * from './hooks';
