+++ src/pages/BondsPage/components/BondModal/BondModal.module.scss
@@ -54,28 +54,6 @@
   font: var(--font-text-6);
 }
 
-body .input {
-  height: 48px;
-
-  @media screen and (max-width: 1024px) {
-    input {
-      height: 40px;
-    }
-
-    height: 40px;
-  }
-}
-
-.btn {
-  margin-top: 16px;
-  height: 48px;
-  width: 100%;
-
-  @media screen and (max-width: 767px) {
-    height: 40px;
-  }
-}
-
 .closeModalSection {
   height: 40px;
   display: flex;
@@ -129,39 +107,5 @@ body .input {
 
   @media screen and (max-width: 1024px) {
     width: 10px;
-  }
-}
-.infoWrapper {
-  display: flex;
-  flex-direction: column;
-  gap: 9.5px;
-  margin-top: 50px;
-}
-
-.info {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.infoTitle {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  color: var(--secondary-text);
-  font: var(--font-text-6);
-  text-transform: uppercase;
-}
-
-.infoValue {
-  color: var(--secondary-border);
-  font: var(--font-text-8);
-}
-
-.questionIcon {
-  svg {
-    fill: var(--blue-color);
-    width: 18px;
-    height: 18px;
   }
 }
