+++ src/pages/BondsPage/components/BondModal/BondModal.tsx
@@ -1,22 +1,9 @@
 import { FC } from 'react';
-import { SOL_TOKEN } from '@frakt-protocol/frakt-sdk';
-import { Controller } from 'react-hook-form';
 
-import { Modal } from '../../../../components/Modal';
-import { TokenFieldWithBalance } from '../../../../components/TokenField';
-import ChangeSidesButton from '../ChangeSidesButton';
-import styles from './BondModal.module.scss';
-import { InputControlsNames, useBondModal } from './hooks';
 import { CloseModalIcon, SolanaIcon } from '../../../../icons';
-import Button from '../../../../components/Button';
-import SettingsModal from './SettingsModal';
-import {
-  getSolBalanceValue,
-  getCorrectSolWalletBalance,
-} from '../../../../utils';
-import { useNativeAccount } from '../../../../utils/accounts';
-import { QuestionCircleOutlined } from '@ant-design/icons';
-import Tooltip from '../../../../components/Tooltip';
+import SwapForm from '../../../../componentsNew/SwapModal';
+import { Modal } from '../../../../components/Modal';
+import styles from './BondModal.module.scss';
 
 interface BondModalProps {
   visible: boolean;
@@ -27,19 +14,6 @@ interface BondModalProps {
 }
 
 const BondModal: FC<BondModalProps> = ({ visible, onCancel }) => {
-  const {
-    control,
-    payToken,
-    receiveToken,
-    slippage,
-    setSlippage,
-    changeSides,
-  } = useBondModal();
-
-  const { account } = useNativeAccount();
-  const solBalanceValue = getSolBalanceValue(account);
-  const solWalletBalance = getCorrectSolWalletBalance(solBalanceValue);
-
   return (
     <Modal
       visible={visible}
@@ -73,85 +47,7 @@ const BondModal: FC<BondModalProps> = ({ visible, onCancel }) => {
           <p className={styles.title}>Roi</p>
         </div>
       </div>
-
-      <SettingsModal slippage={slippage} setSlippage={setSlippage} />
-      <Controller
-        control={control}
-        name={InputControlsNames.PAY_VALUE}
-        render={({ field: { onChange, value } }) => (
-          <TokenFieldWithBalance
-            className={styles.input}
-            value={value}
-            onValueChange={onChange}
-            tokensList={[SOL_TOKEN]}
-            currentToken={payToken}
-            label="Pay"
-            lpBalance={Number(solWalletBalance)}
-            showMaxButton
-          />
-        )}
-      />
-      <ChangeSidesButton onClick={changeSides} />
-      <Controller
-        control={control}
-        name={InputControlsNames.RECEIVE_VALUE}
-        render={({ field: { onChange, value } }) => (
-          <TokenFieldWithBalance
-            className={styles.input}
-            value={value}
-            onValueChange={onChange}
-            currentToken={receiveToken}
-            tokensList={[SOL_TOKEN]}
-            label="Receive"
-            disabled
-          />
-        )}
-      />
-      <div className={styles.infoWrapper}>
-        <div className={styles.info}>
-          <span className={styles.infoTitle}>
-            Slippage Tolerance
-            <Tooltip
-              placement="top"
-              trigger="hover"
-              overlay="The maximum difference between your estimated price and execution price."
-            >
-              <QuestionCircleOutlined className={styles.questionIcon} />
-            </Tooltip>
-          </span>
-          <span className={styles.infoValue}>{`${slippage}%`}</span>
-        </div>
-        <div className={styles.info}>
-          <span className={styles.infoTitle}>
-            Minimum Received
-            <Tooltip
-              placement="top"
-              trigger="hover"
-              overlay="The least amount of tokens you will recieve on this trade"
-            >
-              <QuestionCircleOutlined className={styles.questionIcon} />
-            </Tooltip>
-          </span>
-          <span className={styles.infoValue}>22 $SMB</span>
-        </div>
-        <div className={styles.info}>
-          <span className={styles.infoTitle}>
-            Price Impact
-            <Tooltip
-              placement="top"
-              trigger="hover"
-              overlay="The difference between the market price and estimated price due to trade size"
-            >
-              <QuestionCircleOutlined className={styles.questionIcon} />
-            </Tooltip>
-          </span>
-          <span className={styles.infoValue}>100.00 %</span>
-        </div>
-      </div>
-
-      <Button className={styles.btn} type="secondary">
-        Swap
-      </Button>
+      <SwapForm />
     </Modal>
   );
 };
