--- src/pages/BondsPage/components/BondModal/SettingsModal/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SettingsModal';
