--- src/pages/BorrowPage/components/ShortTermFields/ShortTermFields.module.scss
@@ -1,82 +1,0 @@
-.fieldWrapper {
-  display: flex;
-  flex-direction: column;
-}
-
-.fieldDesc {
-  color: var(--secondary-border);
-  font: var(--font-text-8);
-  font-size: 14px;
-  padding-top: 8px;
-  text-align: center;
-}
-
-.valueFieldValue {
-  margin-right: 5px;
-}
-
-.staticValue {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 0 10px;
-
-  color: var(--secondary-border);
-  font: var(--font-text-5);
-  text-transform: uppercase;
-
-  // &::after {
-  //   content: '';
-  //   flex-grow: 1;
-  //   height: 1px;
-  //   display: block;
-  //   background: var(--color-gray-4);
-  //   order: 2;
-  // }
-}
-
-.tooltipWrapper {
-  display: flex;
-  flex-direction: row-reverse;
-  align-items: center;
-  gap: 5px;
-  order: 3;
-}
-
-.questionIcon {
-  margin-top: -2px;
-  svg {
-    fill: var(--blue-color);
-    width: 18px;
-    height: 18px;
-  }
-}
-
-.staticValueAccent {
-  color: var(--secondary-border);
-}
-
-.staticValueWrapper {
-  display: flex;
-  flex-direction: column;
-  padding-top: 20px;
-  gap: 8px;
-}
-
-.staticValueTitle {
-  color: var(--secondary-text);
-  font: var(--font-text-6);
-  font-size: 12px;
-  order: 1;
-  max-width: 100px;
-
-  @media (max-width: 767px) {
-    font-size: 10px;
-    max-width: unset;
-  }
-}
-.staticValueData {
-  color: var(--secondary-border);
-  font: var(--font-text-8);
-  order: 3;
-}
